requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b0
boto3==1.17.69
botocore==1.20.69
click==7.1.2
Django==3.2
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-pdb==0.6.2
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
